typecheck adapters
