+++ packages/adapter-cloudflare/package.json
@@ -25,13 +25,14 @@
 	"scripts": {
 		"lint": "eslint --ignore-path .gitignore \"**/*.{ts,js,svelte}\" && npm run check-format",
 		"format": "npm run check-format -- --write",
+		"check": "tsc",
 		"check-format": "prettier --check . --config ../../.prettierrc --ignore-path .gitignore"
 	},
 	"dependencies": {
 		"esbuild": "^0.14.21"
 	},
 	"devDependencies": {
-		"@sveltejs/kit": "workspace:*"
+		"typescript": "^4.6.2"
 	},
 	"publishConfig": {
 		"access": "public"
