+++ packages/adapter-netlify/package.json
@@ -25,6 +25,7 @@
 		"dev": "rimraf files && rollup -cw",
 		"build": "rimraf files && rollup -c",
 		"test": "uvu src \"(spec\\.js|test[\\\\/]index\\.js)\"",
+		"check": "tsc",
 		"lint": "eslint --ignore-path .gitignore \"**/*.{ts,js,svelte}\" && npm run check-format",
 		"format": "npm run check-format -- --write",
 		"check-format": "prettier --check . --config ../../.prettierrc --ignore-path .gitignore",
@@ -43,6 +44,7 @@
 		"@sveltejs/kit": "workspace:*",
 		"rimraf": "^3.0.2",
 		"rollup": "^2.58.0",
+		"typescript": "^4.6.2",
 		"uvu": "^0.5.2"
 	}
 }
