+++ packages/adapter-netlify/tsconfig.json
@@ -4,10 +4,13 @@
 		"checkJs": true,
 		"noEmit": true,
 		"noImplicitAny": true,
-		"target": "es2020",
 		"module": "es2020",
 		"moduleResolution": "node",
-		"allowSyntheticDefaultImports": true
+		"allowSyntheticDefaultImports": true,
+		"baseUrl": ".",
+		"paths": {
+			"@sveltejs/kit": ["../kit/types/index"]
+		}
 	},
-	"include": ["./index.js", "src"]
+	"include": ["*.js", "src/**/*.js", "ambient.d.ts"]
 }
