+++ packages/adapter-node/ambient.d.ts
