+++ packages/adapter-node/index.d.ts
@@ -1,5 +1,11 @@
 import { Adapter } from '@sveltejs/kit';
 
+declare global {
+	const HOST_ENV: string;
+	const PATH_ENV: string;
+	const PORT_ENV: string;
+}
+
 interface AdapterOptions {
 	out?: string;
 	precompress?: boolean;
