+++ packages/adapter-node/src/handler.d.ts
@@ -1,0 +1,9 @@
+import type { Handle } from '@sveltejs/kit';
+
+declare global {
+	const ORIGIN: string;
+	const HOST_HEADER: string;
+	const PROTOCOL_HEADER: string;
+}
+
+export const handler: Handle;
