+++ packages/adapter-static/package.json
@@ -34,6 +34,7 @@
 		"port-authority": "^1.1.2",
 		"sirv": "^2.0.0",
 		"svelte": "^3.44.2",
+		"typescript": "^4.6.2",
 		"uvu": "^0.5.2"
 	}
 }
