+++ packages/adapter-vercel/ambient.d.ts
